requirements.txt
# requirements for running agenda
dj-database-url==0.4.0
Django==1.9.7
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
djangorestframework

